requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
dj-database-url==3.0.1
Django==6.0
djangorestframework==3.16.1
filelock==3.20.1
fsspec==2025.12.0
gunicorn==23.0.0
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.46.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.1
pooch==1.8.2
PyMatting==1.1.14
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
rembg==2.0.69
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.16.3
setuptools==70.2.0
six==1.17.0
sqlparse==0.5.4
sympy==1.14.0
tifffile==2025.12.12
tokenizers==0.22.1
torch==2.9.1+cpu
torchaudio==2.9.1+cpu
torchvision==0.24.1+cpu
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
whitenoise==6.11.0
